Title: 십진수와 이진수, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 241017/십진수와 이진수/decimal-and-binary-number.java
@@ -3,8 +3,8 @@ import java.io.*;
 public class Main {
     public static void main(String[] args) throws IOException{
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        int num = Integer.parseInt(br.readLine(),2)*17;
-        System.out.println(Integer.toBinaryString(num));
+        long num = Integer.parseInt(br.readLine(),2)*17;
+        System.out.println(Long.toBinaryString(num));
         
     }
 }
